Separated the calculator into a separate component from Home page for testability
